+++ e2e/test-env/src/client/RicosTestApp/RicosTestApp.jsx
@@ -9,11 +9,7 @@ import { default as viewerPlugins } from './viewerPlugins';
 import './styles.global.scss';
 import theme from '../../../../../examples/main/shared/theme/theme';
 import { testVideos } from '../../../../../examples/main/shared/utils/mock';
-import {
-  TextSelectionToolbar,
-  ViewerInlineToolBar,
-  TwitterButton,
-} from 'wix-rich-content-text-selection-toolbar';
+import { TextSelectionToolbar, TwitterButton } from 'wix-rich-content-text-selection-toolbar';
 
 const onVideoSelected = (url, updateEntity) => {
   setTimeout(() => updateEntity(testVideos[1]), 1);
@@ -68,7 +64,7 @@ class RicosTestApp extends PureComponent {
           cssOverride={theme}
           seoSettings={seoMode}
         />
-        <TextSelectionToolbar container={this.viewerRef.current} ToolBar={ViewerInlineToolBar}>
+        <TextSelectionToolbar container={this.viewerRef.current}>
           {selectedText => <TwitterButton selectedText={selectedText} />}
         </TextSelectionToolbar>
       </>