+++ examples/main/shared/viewer/Viewer.jsx
@@ -8,11 +8,7 @@ import getImagesData from 'wix-rich-content-fullscreen/dist/lib/getImagesData';
 import Fullscreen from 'wix-rich-content-fullscreen';
 import 'wix-rich-content-fullscreen/dist/styles.min.css';
 import { IMAGE_TYPE } from 'wix-rich-content-plugin-image/dist/module.viewer';
-import {
-  TextSelectionToolbar,
-  ViewerInlineToolBar,
-  TwitterButton,
-} from 'wix-rich-content-text-selection-toolbar';
+import { TextSelectionToolbar, TwitterButton } from 'wix-rich-content-text-selection-toolbar';
 import { GALLERY_TYPE } from 'wix-rich-content-plugin-gallery';
 const anchorTarget = '_top';
 const relValue = 'noreferrer';
@@ -90,7 +86,7 @@ export default class Viewer extends PureComponent {
             />
           )}
           {!isMobile ? (
-            <TextSelectionToolbar container={this.viewerRef.current} ToolBar={ViewerInlineToolBar}>
+            <TextSelectionToolbar container={this.viewerRef.current}>
               {selectedText => <TwitterButton selectedText={selectedText} />}
             </TextSelectionToolbar>
           ) : null}