+++ packages/text-selection-toolbar/web/src/TextSelectionToolbar.jsx
@@ -1,5 +1,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
+import styles from '../statics/styles/viewer-inline-toolbar.rtlignore.scss';
+import addTextSelectionListener from './TextSelectionListener';
 import { debounce } from 'lodash';
 
 export default class TextSelectionToolbar extends React.Component {
@@ -8,58 +10,51 @@ export default class TextSelectionToolbar extends React.Component {
     this.state = { selectedText: '' };
   }
 
-  getSelectedText = selection => selection.toString().replace(/(\r\n|\r|\n){2,}/g, ' ');
-
-  getSelectionPosition = selection => {
-    const parent = selection.anchorNode.parentNode;
-    const parentRect = parent.getBoundingClientRect();
-    const selectionRect = selection.getRangeAt(0).getBoundingClientRect();
-    const parentTop = parent.offsetTop;
-    const selectionOffesetFromParent = selectionRect.top - parentRect.top;
-    const y = parentTop + selectionOffesetFromParent;
-    const { x, width } = selectionRect;
-    return { x: x + width / 2, y };
-  };
-
-  handleSelection = debounce(() => {
-    const { container } = this.props;
-    const selection = document.getSelection();
-    const selectionElement = selection.anchorNode?.parentElement;
-    let text = null;
-    let position = null;
-    if (
-      selection.rangeCount > 0 &&
-      (container.contains(selectionElement) || selectionElement === container)
-    ) {
-      text = this.getSelectedText(selection);
-      position = this.getSelectionPosition(selection);
-    }
-    this.setState({ selectedText: text, position });
-  }, 100);
-
   componentDidMount() {
-    document.addEventListener('selectionchange', this.handleSelection);
+    this.addTextSelectionListener(this.props.container);
   }
 
   componentWillUnmount() {
-    document.removeEventListener('selectionchange', this.handleSelection, false);
+    this?.removeTextSelectionListener();
   }
 
+  componentWillReceiveProps(nextPros) {
+    this.addTextSelectionListener(nextPros.container);
+  }
+
+  addTextSelectionListener = container => {
+    if (!this.removeTextSelectionListene && container) {
+      this.removeTextSelectionListener = addTextSelectionListener(container, this.setSelectedText);
+    }
+  };
+
+  setSelectedText = debounce(
+    (selectedText, selectedTextPosition) => this.setState({ selectedText, selectedTextPosition }),
+    50
+  );
+
   render() {
-    const { ToolBar, container, children } = this.props;
-    const { selectedText, position } = this.state;
+    const { selectedText, selectedTextPosition } = this.state;
+    const { container, children } = this.props;
+    if (!selectedText || !container) {
+      return null;
+    }
+    const { left } = container.getBoundingClientRect();
+    const style = {
+      top: selectedTextPosition.y,
+      left: selectedTextPosition.x - left,
+    };
+
     return (
-      selectedText && (
-        <ToolBar position={position} container={container}>
-          {children(selectedText)}
-        </ToolBar>
-      )
+      <div className={styles.toolbar} style={style}>
+        {children(selectedText)}
+      </div>
     );
   }
 }
 
 TextSelectionToolbar.propTypes = {
-  ToolBar: PropTypes.any.isRequired,
+  position: PropTypes.object.isRequired,
   children: PropTypes.any,
   container: PropTypes.object.isRequired,
 };