+++ packages/text-selection-toolbar/web/src/index.ts
@@ -1,5 +1,4 @@
 import TextSelectionToolbar from './TextSelectionToolbar.jsx';
-import ViewerInlineToolBar from './ViewerInlineToolBar.jsx';
 import TwitterButton from './TwitterButton.jsx';
 
-export { TextSelectionToolbar, ViewerInlineToolBar, TwitterButton };
+export { TextSelectionToolbar, TwitterButton };